# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger",
        "description": "API CRUD Books",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8040"
        }
    ],
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get paginated list of books",
                "description": "Returns a paginated list of books",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Number of books per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookResource"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "format": "url"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "format": "url"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "format": "url",
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "format": "url"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "format": "url",
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Store new book details",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the book",
                                        "type": "string"
                                    },
                                    "author": {
                                        "description": "Author of the book",
                                        "type": "string"
                                    },
                                    "isbn": {
                                        "description": "ISBN of the book",
                                        "type": "string"
                                    },
                                    "publication_date": {
                                        "description": "Publication Date of the book",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "pages": {
                                        "description": "Pages of the book",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get book information",
                "description": "Get book details by ID",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update book information",
                "description": "Update book details by ID",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the book",
                                        "type": "string"
                                    },
                                    "author": {
                                        "description": "Author of the book",
                                        "type": "string"
                                    },
                                    "isbn": {
                                        "description": "ISBN of the book",
                                        "type": "string"
                                    },
                                    "publication_date": {
                                        "description": "Publication Date of the book",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "pages": {
                                        "description": "Pages of the book",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete book record",
                "description": "Delete book by ID",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookResource": {
                "properties": {
                    "id": {
                        "description": "Id of the book",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string"
                    },
                    "author": {
                        "description": "Author of the book",
                        "type": "string"
                    },
                    "isbn": {
                        "description": "ISBN of the book",
                        "type": "string"
                    },
                    "publication_date": {
                        "description": "Publication Date of the book",
                        "type": "string",
                        "format": "date"
                    },
                    "pages": {
                        "description": "Pages of the book",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}